Extract VerifyKernelFuncs into pkg/ebpf/ksyms leaf package
VerifyKernelFuncs and GetKernelSymbolsAddressesNoCache are pure
/proc/kallsyms parsers — they have no dependency on the rest of pkg/ebpf
(BTF, manager, perf, CO-RE, etc.). Today the 10 callers that only need
VerifyKernelFuncs (kernelbugs, modifiers, several tracer subpackages,
USM, and the security probe) pull in all of pkg/ebpf as a transitive
dep just for these helpers.

Move them into a new leaf package pkg/ebpf/ksyms with no dependency on
pkg/ebpf. ksyms_bpf.go stays in pkg/ebpf (it uses LoadCOREAsset) and
delegates to the new package.

This also clears the way for an upcoming PR that adds a pkg/ebpf/features
import to pkg/ebpf/ebpftest. As of today, that import would close a
test-time cycle:

  pkg/ebpf (tests) -> ebpftest -> features -> kernelbugs -> pkg/ebpf

With kernelbugs depending on the new leaf package instead of pkg/ebpf,
the cycle breaks at its origin.
