+++ pkg/ebpf/BUILD.bazel
@@ -180,7 +180,6 @@ go_library(
         "debug_handlers.go",
         "debuglog_dumper.go",
         "ebpf.go",
-        "ksyms.go",
         "ksyms_bpf.go",
         "lockcontention.go",
         "lockcontention_noop.go",
@@ -209,9 +208,9 @@ go_library(
         "//pkg/config/setup",
         "//pkg/config/utils",
         "//pkg/ebpf/bytecode",
+        "//pkg/ebpf/ksyms",
         "//pkg/ebpf/maps",
         "//pkg/ebpf/names",
-        "//pkg/process/util",
         "//pkg/remoteconfig/state",
         "//pkg/system-probe/config",
         "//pkg/util/archive",
@@ -263,7 +262,6 @@ go_test(
         "co_re_telemetry_test.go",
         "feature_test.go",
         "ksyms_bpf_test.go",
-        "ksyms_test.go",
         "lockcontention_test.go",
         "manager_test.go",
         "map_cleaner_test.go",
