+++ pkg/ebpf/ksyms/testdata/kallsyms.supported
