+++ pkg/network/tracer/connection/util/kernel.go
@@ -8,7 +8,7 @@
 package util
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/ebpf"
+	"github.com/DataDog/datadog-agent/pkg/ebpf/ksyms"
 	"github.com/DataDog/datadog-agent/pkg/util/kernel"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -19,7 +19,7 @@ import (
 // this change into 6.4.0 to pick up a CVE so the version number is not reliable.
 // Instead, we directly check if the function exists.
 func HasTCPSendPage(kv kernel.Version) bool {
-	missing, err := ebpf.VerifyKernelFuncs("tcp_sendpage")
+	missing, err := ksyms.VerifyKernelFuncs("tcp_sendpage")
 	if err == nil {
 		return len(missing) == 0
 	}
