+++ pkg/network/usm/sharedlibraries/ebpf.go
@@ -28,6 +28,7 @@ import (
 	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/ebpf/bytecode"
 	bugs "github.com/DataDog/datadog-agent/pkg/ebpf/kernelbugs"
+	"github.com/DataDog/datadog-agent/pkg/ebpf/ksyms"
 	"github.com/DataDog/datadog-agent/pkg/ebpf/modifiers"
 	"github.com/DataDog/datadog-agent/pkg/ebpf/perf"
 	ebpftelemetry "github.com/DataDog/datadog-agent/pkg/ebpf/telemetry"
@@ -506,7 +507,7 @@ func (e *EbpfProgram) initPrebuilt() error {
 }
 
 func sysOpenAt2Supported() bool {
-	missing, err := ddebpf.VerifyKernelFuncs("do_sys_openat2")
+	missing, err := ksyms.VerifyKernelFuncs("do_sys_openat2")
 
 	return err == nil && len(missing) == 0
 }
